requirements_test.txt
pytest>=8.0
pytest-asyncio>=0.24
pytest-homeassistant-custom-component>=0.13.0
pytest-cov>=6.0
